+++ .breakpoints
@@ -1,0 +1,12 @@
+{
+  "files": {
+    "embeds.js": [
+      {
+        "id": "89c2fb54-67a3-4e34-a5c8-5b2dab94376c",
+        "line": 43,
+        "version": 355,
+        "index": 1597
+      }
+    ]
+  }
+}
